+++ EUS/GameplayScene.cpp
@@ -59,7 +59,7 @@ void GameplayScene::draw(const float deltaTime)  {
 	int x, y;
 	SDL_GetMouseState(&x, &y);
 
-	p->findPath(found, path, infos, pmath::Vec2f(x, y), pmath::Vec2f(384.0f, 384.0f));
+	p->findPath(found, path, infos, pmath::Vec2f(static_cast<float>(x), static_cast<float>(y)), pmath::Vec2f(384.0f, 384.0f));
 
 	if (found) {
 		for (size_t i = 0; i < path.size(); i++) {
